[kernel][vm] Fix argument validation when upper limit flag is set.
Fix slight differences in argument validation of the offset and
size params to vmar_allocate, depending on whether the upper limit
flag is set. The offset and size should be able to reach the end of
a VMAR when the upper limit flag is set.

Test: Added tests pass.
Change-Id: I6a64cd349adcc45beb7073066c2bd2de22cc1a2d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/406740
Reviewed-by: Adrian Danis <adanis@google.com>
Testability-Review: Adrian Danis <adanis@google.com>
Commit-Queue: Corey Tabaka <eieio@google.com>