[kernel][debuglog] Move the main Debuglog object to LazyInit
This lets us initialize it earlier and as a result safely print earlier
in the kernel. Rearrange the init order in the top of the kernel and
printf to ensure this doesn't break in the future.

Added some constructor trace code, disabled by default, to help with
printing constructor routines as they're called.

Mark FILE object as constexpr to ensure that the default stdout object
is also ready to go from the start. Would be a good candidate for
constinit when the compiler is eventually switched to c++20.

Tested: booted on x86 and arm qemu with and without
kernel.bypass-debuglog=1. Run through CQ to boot on real hardware.

Change-Id: I84c27e05743b0870dca02ff22d9aee6d552430d2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/406355
Commit-Queue: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: David Greenaway <dgreenaway@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Testability-Review: David Greenaway <dgreenaway@google.com>