[build] Make the build trully independent of the absolute path
-fdebug-prefix-map isn't sufficient to make the build truly independent
of the build absolute path. While it eliminates the absolute path from
the debug info, the absolute path is still a part of the compile command
which is sent to goma, used by tools via compilation database etc. To
address this limitation, Clang provides a -fdebug-compilation-dir option
which doesn't require the absolute path.

Bug: 55606
Change-Id: If5488678c9484940e5dd82c8808cc9c113106756
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/352242
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: David Greenaway <dgreenaway@google.com>
Reviewed-by: George Kulakowski <kulakowski@google.com>
Reviewed-by: Nathan Mulcahey <nmulcahey@google.com>