[ubench] Add core affinity to ContextSwitch.
The ContextSwitch benchmark appears to be set up to measure the context
switching time between two threads on the same core. However it doesn't
specify to the kernel that the threads should be running on the same
core so it is subject to timing noise depending on where the threads end
up being scheduled.

This change uses affinity to attach each threadpair to a single core to
make the timing more stable and the results more accurate.

Test: ran ubench locally and ran perfcomp on bots.
Change-Id: Iff343dd6faebaad8f4e92f5ef4ca85ae29519b4c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405083
Commit-Queue: Ed Coyne <edcoyne@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Mark Seaborn <mseaborn@google.com>
Testability-Review: Christopher Johnson <crjohns@google.com>