[settings] Migrate facade back to watch
This a step in the process to rename Watch2 to Watch. A previous api
change required us to add a version of Watch that didn't return a
result/error, so it had to be renamed Watch2 so that both could
exist for a soft-transition. This change is to migrate back to the name
Watch.

Fixed: 55685
Test: fx test sl4f-service-tests

Change-Id: I24c28dc750b9fc2594db62a6a67d910480264c59
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/406540
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Testability-Review: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Chip Fukuhara <cfukuhara@google.com>