[fx/pkg] deprecate the notion of "push"
The concept of "push" is misleading as users extrapolate that this
means something akin to "install", and make other extrapolations such
as the fileystem underneath a running component could be changed by
this action (which is intentionally not the case).

A new command `fx run` is offered as a replacement for users who want
the one-shot behavior to build and then run a new thing. This command
also introduces a slightly new feature, in that it will attempt to
terminate a previously running instance of the same component.

Change-Id: Ia2917d0fe64ece59b0c3554dbe43c5b689e3e8fe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/400493
Commit-Queue: James Tucker <raggi@google.com>
Reviewed-by: Renato Mangini Dias <mangini@google.com>
Testability-Review: James Tucker <raggi@google.com>