[clang] Temporarily disable -Wsuggest-override
https://reviews.llvm.org/D82728 added this warning which is resulting in
a compile error on the clang canary builders. Disable it for now to
allow a roll, then we should fix instances of these after the roll.

Also remove the the one instance this is used in audio_core. This
warning would not have been exposed in the build until now.

Bug: 56202
Change-Id: I41e1ddbbdf2a03a83e22a7ef166474017a27b45f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/406925
Commit-Queue: Leonard Chan <leonardchan@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>