[Security] Create the security codelab.
This CL includes smart-door, smart-door-memory, codelab and solution
components. The smart-door and smart-door-memory is written
intentionally to contain vulnerabilites.

The smart-door is a very simple fake 'smart door' that allows user
to add home member by using username and password, and allows opening
using the username and password. If none of the username matches, it
would check against a hardcoded admin password to see if it is admin
trying to 'open' the door.

It uses 'smart-door-memory' component also in the codelab as the
persistent storage. The 'smart-door-memory' is a simple token
based storage that allows user to read and write to an isolated
storage space referred by the token.

DO NOT COPY ANY OF THE CODE IN THIS COMPONENT!!!

Tested: Unit tests.
Change-Id: Ia57ca5344ad91e205605e998e5928e456a30f321
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383594
Commit-Queue: Yu Shan <shanyu@google.com>
Testability-Review: Drew Fisher <zarvox@google.com>
Reviewed-by: Bryan Henry <bryanhenry@google.com>
Reviewed-by: Kostya Kortchinsky <kostyak@google.com>