[kernel][autospinlock] Remove assert, add nonnull check
From looking at dissassembly, the assert for non null at the top of
the AutoSpinLock constructor was causing the compiler to generate a
few copies of the function instead of inlining it. Passing a bogus
pointer to is is pretty hard to do and a pretty low risk, so just
replace it with a NONNULL compiler check and enjoy the slightly
better code gen.

No functional change (except removal of an assert).

Tested: built on arm and x86 and validated that the codegen looks nicer.
Change-Id: I8182029d718ff28f3cc5d73c12d0e278017f93c1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/406357
Commit-Queue: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: David Greenaway <dgreenaway@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Testability-Review: David Greenaway <dgreenaway@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>