[display] Remove ZX_DEBUG_ASSERT for amlogic RDMA.
This debug assert is unnecessary because the display hardware protocol
is stateless. The only time the assert can fire is when a client is
sending configs without waiting for vsync, and therefore the missed
vsync notifications caused by failing to apply will not cause problems
for the client connection.

The assertion is replaced with Inspect stats that can be used to spot
this issue in the field.

Fixed: 50650
Test: e2e test on a local Astro

Change-Id: I48ffed7aab124f583f7d543ee10b2e75b6a8120a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/406919
Reviewed-by: Payam Moradshahi <payamm@google.com>
Testability-Review: Payam Moradshahi <payamm@google.com>
Commit-Queue: Randall Bosetti <rlb@google.com>