[bt][bt-gap] Rename and move DiscoveryRequestToken from host dispatcher to host device
Previously, host_dispatcher.rs handled the creating and distribution of
DiscoveryRequestTokens upon Access FIDL StartDiscovery requests. Now we
have encapsulated this behavior to host_device.rs, since a discovery
session is associated with a specific host adapter. This way, a token
cannot be erroneously created/modified/dropped in the dispatcher
without calling into the device API.

This commit is a preliminary step to resolving Issue 640, which is to
make Discovery a system-wide state, over all active adapters. Resolution
of this issue will provide the framework for other states  (such as
Discoverable, the token for which is untouched here and should be moved
to host_device.rs in the future) to become system-wide as well. (i.e.,
if one adapter is dropped, and a new adapter is activated, these states
can persist on the new adapter)

Bug: 640
Change-Id: Ifc87090717061312c518b008340614e40b9e670e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/406391
Commit-Queue: Ali Saeed <saeedali@google.com>
Reviewed-by: Nick Pollard <nickpollard@google.com>
Testability-Review: Nick Pollard <nickpollard@google.com>