[hlcpp][gidl] In HLCPP conformance test, don't fake out transaction header
Motivation:
- This adds a lot of complication to the test which makes it harder
to understand and verify that it is right.
- The decode failure test was actually not working - it was catching
failures due to the wrong FIDL type.
- Need to update the coding table for the transaction header manually
when coding tables change (this is actually my original motivation
because a change I'm working on made it complex to build the coding
table for this manually).

Change-Id: I373c51fc6a56a8648ac59e697266a562a634b9b1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407154
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
Testability-Review: Ian McKellar <ianloic@google.com>