[storage] Fix the client so that it sends CloseFifo when dropped.
Also fixed a bug in the block server where closing the fifo could get
blocked if unable to return responses on the fifo (which can be the case
if the client isn't servicing the fifo).

Changed termination for the Rust client so that the fifo is actively
closed.

Bug: 54733
Change-Id: Id50c2f5cba0260b717f3cf79b724b033846a0c0d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/404198
Reviewed-by: Gianfranco Valentino <gevalentino@google.com>
Commit-Queue: Chris Suter <csuter@google.com>
Testability-Review: Gianfranco Valentino <gevalentino@google.com>