[scrutiny] New dashboard landing page.
This change adds a new landing page for scrutiny with a button that you
click to move to the graph. This enables us to add more visualizers in
the future with a uniformed landing dashboard. This is just a very
incremental change.

Change-Id: If37b0a24f36c364fd10e4c26b64471985bacb568
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407211
Reviewed-by: Daniel Hwang <arkay@google.com>
Testability-Review: Benjamin Wright <benwright@google.com>
Commit-Queue: Benjamin Wright <benwright@google.com>