[virtualization] Disable virtio-magma on Zircon
It looks like a recent change to how the vulkan libraries are built is
causing an issue with virtio-magma, resulting in flakes. Disable
virtio-magma in Zircon (and Debian) to deal with this, since neither
environment requires it.

Fixed: 56119
Change-Id: I6dd3043adac3e58ba59a2d0c2b907bfc42a5886e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407414
Testability-Review: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Alex Legg <alexlegg@google.com>
Commit-Queue: Abdulla Kamar <abdulla@google.com>