[scripts][fuzzing] Handle case where package is in TUF repo but not on disk
`pkgctl pkg-status` returns an exit code of 2 when a package is built as
a universe package but not yet resolved to disk. This was being
interpreted as an error, preventing package resolution from proceeding
as expected. This CL ensures that such an exit code is handled properly,
and extends the process mocking framework to handle variable exit codes
so that the behavior can be tested.

Fixed: 55235
Test: fx fuzz unittest
Change-Id: I46484aac71672d5b29f49a73b95998feab1c588b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/406685
Reviewed-by: Aaron Green <aarongreen@google.com>
Testability-Review: Aaron Green <aarongreen@google.com>
Commit-Queue: Cameron Finucane <eep@google.com>