[debugger] Remove the arch provider
Moves the breakpoint and watchpoint functions to the DebugRegisters
object. The DebugRegisters implementation is now split off into an x64
and an arm64 implementation.

Move WatchpointInstallationResult to WatchpointInfo in its own file.
This is now used for both installs and queries of the DebugRegisters.
The "status" is removed and failure is indicated (when needed) by
returning in a std::optional.

Remove some zx_status_t results from affected functions. Booleans and
std::optional make the API cleaner, and all we were ever doing with the
results was printing the errors, sometimes multiple times. Now we print
it once and just pass a boolean up the stack.

The breakpoint instruction checking function is split out into an
architecture-specific version that checks a passed-in value, and a
variant on the DebuggedThread that actually fetches the memory.

Change-Id: I11118eb7cfa472082bfe37251ceb1a4738de5df7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407283
Reviewed-by: Dangyi Liu <dangyi@google.com>
Testability-Review: Brett Wilson <brettw@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>