[feedback] Add test FIDL to query CrashReporter
CrashReporter, like Cobalt, is a component that is sent data and
is often ignored when it responds. This make integration tests
challenging because a test may not fail if something is wrong with
a component's crash reporting logic and a crash report is not filed.
Adding a query interface allows component authors to easily check
that crash reports are filed in integration tests.

Bug: 55875

Change-Id: Ie497203c801f73a9c8511d93a42be64e6387d9f5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407118
Commit-Queue: Alex Pankhurst <pankhurst@google.com>
Reviewed-by: Francois Rousseau <frousseau@google.com>
Testability-Review: Francois Rousseau <frousseau@google.com>
Testability-Review: Alex Pankhurst <pankhurst@google.com>