[bt][a2dp] Refactor RtpPacketBuilder into library
Delegate the building of the RtpPacketBuilder to the MediaCodecConfig,
as it holds all the information to build it.

Introduce the RtpPacketBuilder trait which will be used in the future
to build more codec-specific behavior for RTP packet builders.

Bug: 52616

Test: refactor, tests moved to bt-a2dp-tests

Multiply: bt-a2dp-tests
Change-Id: I1313f396828dd25b3f805e9956c02a704daa0a0f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/406319
Commit-Queue: Marie Janssen 💖 <jamuraa@google.com>
Testability-Review: Marie Janssen 💖 <jamuraa@google.com>
Reviewed-by: Jeff Belgum <belgum@google.com>