[bt][avrcp-tg] Fix log spam from notifications
In AVRCP, notifications are a one-use object. When the notification
value changes, the changed value is sent over the responder, and the
notification object is dropped. However, there are some cases where the
notification value never changes. In these cases, when the Notification
object is dropped, the current value is sent over the responder. The
sending of this value was being logged, and in some cases, the send
fails - this is OK. It usually means the client stopped caring about the
notification value or the client disconnected. This CL removes any
logging around the send operation as this was causing log spam.

While we're here, update the component to use log crate macros - this is
the standard guidance to avoid mismatches in severities across
components.

Bug: 56335
Test: Log change, fx test bt-avrcp-target-tests
Change-Id: Idd1c2e34d2089d41243cc18f121121dd141fd725
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407616
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Marie Janssen 💖 <jamuraa@google.com>
Testability-Review: Marie Janssen 💖 <jamuraa@google.com>