[media][audio_core] Apply OutputPipeline Updates
To continue enabling the AudioTuner service, this CL implements the
proper update of the OutputPipeline with a provided PipelineConfig and
VolumeCurve.

Two key factors to proper replacement of the OutputPipeline are:
	> Maintenance of input streams
	> Execution on the mix thread

Input streams are maintained by forcing the re-route of these inputs
through the new pipeline by removing the output device from the route
graph and adding it following the pipeline update; this takes advantage
of existing mechanisms for routing streams upon the addition of a
device. Because of this, the AudioTuner method to replace the
OutputPipeline is only valid on a device without links.

The actual replacement of the OutputPipeline is executed on the mix
thread so as to align with existing mix tasks and provide a smooth
transition between OutputPipeline configurations. Executing on the mix
thread ensures the application of the new pipeline occurs between mixes,
never during a mix. A 'routable' boolean is used to protect the
OutputPipeline during update; 'routable' is false during update such
that the device plug state cannot change and another update cannot be
simultaneously applied; 'routable' is made true upon update of the
PipelineConfig.

Bug: 52962

Test: fx test audio_pipeline_tests

Change-Id: I4779ad4e5f598bde3245890c5ae70617b40cdfdb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/398960
Commit-Queue: Payton Bradsky <pbradsky@google.com>
Testability-Review: Tim Detwiler <tjdetwiler@google.com>
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>