[bt][a2dp] Migrate AvdtpController to use a2dp::Peer
A2dpController (renamed from AvdtpController) now operates on an
a2dp::Peer, allowing for manipulation of the underlying avdtp::Peer and
the maintanence of state.

Some PTS tests required stateful behavior (i.e
out of band suspend results in the command being sent AND the underlying
Stream being in the right state). Using the a2dp::Peer, which maintains
the collection of Streams, the state of the peer, and the avdtp::Peer
allows us to achieve this.

Updated logging to use log as opposed to fuchsia_syslog.

Updated the A2dpControllerPool to use the already-discovered streams
from the a2dp::Peer, as opposed to attempting to discover when needed.
This was confusing PTS as stray avdtp.discover() commands were being
sent.

Bug: 56032
Test: fx test bt-a2dp-tests, PTS AVDTP/SRC/INT/SIG/SMG/BV-21-C

Change-Id: I88cc6e11dcad6e1232352d9ddd5a05aa473869b7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/406059
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Marie Janssen 💖 <jamuraa@google.com>
Testability-Review: Marie Janssen 💖 <jamuraa@google.com>