[blobfs] Fix race in unmount
Currently, async::Loop requires that we only add or remove paged VMOs
from its internal linked list serially.

Blobfs violates this invariant during teardown. The pager object is
destroyed in the main thread, which iterates through and cancels all of
the paged VMOs in the async::Loop. But it is possible for another thread
(in particular, the journal thread) to simultaneously modify the list,
by deleting a Blob object, which unregisters a paged VMO from the
async::Loop.

This fix is a workaround that makes the race condition impossible, by
ensuring that the journal operations finish before deleting the pager
object. It is a brittle fix and more attention should be given to how to
properly fix issues of this class.

Bug: 55829
Change-Id: Icb060a089f0f25359c7a539f184b25362a8e9934
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407599
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Xyan Bhatnagar <xbhatnag@google.com>
Testability-Review: Rasha Eqbal <rashaeqbal@google.com>