[ninja] Make go/build.py use py3
This is necessary as otherwise when creating the symlink tree, Python 2
will truncate the mtimes of the symlinks, resulting in
older-than-expected mtimes. This is fixed in Py3, and it is hoped this
will resolve the fail no-op as in
https://ci.chromium.org/raw/build/logs.chromium.org/fuchsia/led/fuchsia-try-builder_fuchsia-infra.iam.gserviceaccount.com/ef71493e6e5c8e1586c2d2af0a34f3ec604c4ba2eb663d90e34a591ec361b438/+/build.proto?server=chromium-swarm.appspot.com

    """
    ninja explain: output host_x64/gen/tools/testing/affectedtests/testdata older than most recent input ../../tools/testing/affectedtests/testdata (1594856370119459000 vs 1594856370119459941)
    """

In particular note that the dates are the same except the last three
digits which have truncated nanoseconds to microseconds.

Bug: 56376
Bug: 36581
Change-Id: I25f194a563300e03b5a5637274790dea6da039b7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407784
Commit-Queue: Scott Graham <scottmg@google.com>
Reviewed-by: Randall Bosetti <rlb@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>
Testability-Review: Randall Bosetti <rlb@google.com>