[ninja][build] Ensure there's always a .sym output
When there's no buildid for the kernel, a special case to handle
dump_syms results in skipping emitting the .sym file (by not running
dump_syms). This causes the build to always be dirty because the
'outputs' still specify that the .sym will be output. To avoid this,
touch to make an empty file.

Bug: 56377
Bug: 56374
Bug: 36581
Change-Id: I1283ddd13fa4e31e3d69eadf37f63a5a423204ed
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407777
Commit-Queue: Scott Graham <scottmg@google.com>
Reviewed-by: Francois Rousseau <frousseau@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>