[audio_core] AudioClock refinement
AudioClock objects are passed by reference.These (and their
underlying zx::clock objects) are owned by either an AudioDriver,
BaseRenderer or BaseCapturer, so we can guarantee that the
ClockReference object will outlive those that are given a
reference.

The zx::clock object at the core of AudioClock is now entirely
encapsulated, as shown by the deletion of the clock() accessor
and the addition of utility methods that previously would have
required the raw zx::clock.

AudioClock is updated to include a source (client, device) and a
type (adjustable, non-adjustable, invalid) as well as a clock
domain and a flag that represents whether a client clock has
been designated as controlling the clock hardware.
of 'adjustable' is dropped.

Also, MixStage::SetupMix is reorganized to prepare for subsequent
CL that adds clock reconciliation.

Test: 2 new AudioClock unittests, >50 existing tests adjusted
Bug: 46601

Change-Id: I3078509d39f35bf5c9498e626a1aceb7bd755f9f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405594
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Testability-Review: Tim Detwiler <tjdetwiler@google.com>