[debugger] Fix failed line lookup.
The issue arises when two functions are exactly adjacent in memory. The
line table will contain an "EndSequence" row for the end of the first
function to mark where the previous line table entry ends. But this is a
non-inclusive end marker and will be at the beginning of the following
function.

Previously, when looking up the line information for the beginning of
the second function, we would find the EndSequence marker for the
previous one.

This happens much more often on ARM because the consistent-length
functions always pick up at the next instruction. x86 code often has
padding which will hide this bug.

This also clarifies (but does not change) the behavior of the
LargestLessOrEqual helper function when it encounters a sequence of
equal items, and and adds a test for this case.

Change-Id: If40bce68c95da5e6e3169b73e080785c7fa534ed
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407964
Reviewed-by: Dangyi Liu <dangyi@google.com>
Testability-Review: Brett Wilson <brettw@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>