[kernel][vm] Avoid race in child deletion
This fixes a bug where VmObject referred to in a children_list_ would
be used after it had been destroyed. The use after free occurs as the
children_list_ stores raw pointers, which is mostly fine as it is known
that the child will stay alive until it can acquire the vmo lock. To
avoid a lock ordering inversion OnUserChildRemoved drops the vmo lock
prior to using the object, which is invalid as the 'this' pointer was
a raw pointer from the children_list_.

The solution here is to try and hold a refptr to the child, and only
if a refptr can be upgraded perform the OnUserChildRemoved step so
that the lock can be safely released.

Bug: 56137

Change-Id: I901745fd41cc43d5daa4bddbf3a1452655dda142
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/406757
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Testability-Review: Rasha Eqbal <rashaeqbal@google.com>