[fs_test] Fix AttrTest.WriteSetsModificationTime
Previously, the test would only assert that the modification time of the
file hadn't gone backwards. Fix this by sleeping for the timestamp
granularity of the filesystem and then writing to it, which means we can
assert that the modification time has actually moved forward.

Also add a test that the modification time is updated even if the file
is not closed.

Change-Id: Idb308f855b76ba0b964a8a64aa7b6e772ccc34b5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407516
Commit-Queue: Simon Shields <simonshields@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Testability-Review: Chris Suter <csuter@google.com>