[components] Make event types more consistent and improve tests
Prior to this patch, there were two types of v2 component events: ones
that had payloads and ones that didn't. The ones that didn't had a
different test API than the ones that had payloads. That added a bit to
the confusion of using these events. With this CL, all events have
payloads, if only empty payloads. This allows all events to have a
Result type. This also simplifies the macros in events.rs significantly.

This CL also does a couple of things to improve events integration
tests. It moves the tests in a direction where the test function no
longer has a return value and instead all operations call .unwrap().
This will cause them to crash if the call returns an error. Furthermore,
nested components get syslog and LogSink in order to output the result of the
failed unwrap. These will make debugging tests in the future much easier
as we'll know the exact line number failing and the error description
causing the failure in tests.

Change-Id: I1330baae2ea77739f951ee90c33d5e15e652c28f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407202
Commit-Queue: Fady Samuel <fsamuel@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Xyan Bhatnagar <xbhatnag@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
API-Review: Gary Bressler <geb@google.com>
Testability-Review: Miguel Flores <miguelfrde@google.com>