[tefmocheck] Add check for failing to run bugreport
One of the more common unattributed task failure modes in the last few
days is the bugreport ssh command failing with no exit code when
testrunner executes it at the end of a task.

Fixing that might be as simple as adding retries, but without detection
in place we won't be able to tell if a fix actually works, so adding a
tefmocheck check is a necessary first step either way.

Bug: 56651
Change-Id: Icbb73ab4bd2b1d02810917474ea51b16036a7c08
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/408494
Reviewed-by: Ina Huh <ihuh@google.com>
Testability-Review: Ina Huh <ihuh@google.com>
Commit-Queue: Oliver Newman <olivernewman@google.com>