[paver][vmo] Assert that payload VMO is a clone if pager-backed
Add an assert in PartitionPave to verify that if the payload VMO is
pager-backed, it is also a clone. This guarantees that once its pages
are committed, they will remain in memory and not get evicted. We commit
the payload's pages before writing out a partition to prevent deadlocks
in the block server (due to potential page faults).

Test: Manually verified that `fx update` succeeds with and without
demand paging.

Bug: 48145
Change-Id: Ibbeaa89223448b2deb81d7511ed620a8c2b2b6f5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/408057
Reviewed-by: Adrian Danis <adanis@google.com>
Testability-Review: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>