[serve-remote] supress set-device warnings
I added a feature to set-device that warns the user about running it
under a remote workflow, but this warning fires on subsequent
invocations of serve-remote, which I did not notice because the
remote was behind the version that was being implemented and tested
locally when the warnings were added.

A simple environment variable flag is used to suppress the behavior
for the invocation in question.

Change-Id: Ia0091103a81e1a8681392ff60bb1c222347e84e4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/408095
Reviewed-by: Renato Mangini Dias <mangini@google.com>
Testability-Review: Renato Mangini Dias <mangini@google.com>
Commit-Queue: James Tucker <raggi@google.com>