[kernel][mp][thread] Assert when blocking from an inactive CPU
When a CPU blocks, the scheduler selects another thread to run on the
CPU.  If there are no other threads ready to run, the special idle
thread will be selected.  In the general case, a CPU must be marked
active in order to schedule a thread.  Therefore, a precondition of
blocking is that the calling CPU be marked active.

Mutex acquisition and wait queue entry already assert that it's OK to
block (via arch_blocking_disallowed).  However, the active/inactive
state is not taken into account.

This change sets "blocking disallowed" to true when a CPU transitions
from active to inactive and false when transitioning back.

Additionally, this change sets blocking disallowed to true early on in
the boot sequence of secondary CPUs as they are not yet active.
However, the boot CPU continues to run with blocking allowed prior to
becoming active.  It's OK for the boot CPU to run like this because
the boot CPU will become active before any secondary CPUs are brought
online and before the idle thread (or any other thread) begins
execution.

Tested by booting qemu/x64+kvm repeated for several hours.

Bug: 55862
Change-Id: I63d41b6245a34e35c19bcdc5d56c2ec326f4f12a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/405896
Commit-Queue: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Testability-Review: Corey Tabaka <eieio@google.com>