[netstack] Avoid proxy storage race
Store the event proxy before starting the handler goroutine, avoiding a
race in which the handler goroutine starts and exits entirely before the
proxy is stored, resulting in the proxy being stored indefinitely.

Avoid suppressing ErrBadhandle while I'm here, since it should not be
possible for the handle to be invalid before the handler goroutine is
started.

There is still a small race possible between the handler goroutine's
cleanup and an interface change event, but such logs are expected to
appear only once, and we'd like to be made aware if this occurs.

Fixed: 56733
Change-Id: I7b59780b98684f0443eaaffb566453e8020c5a0e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/408858
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Testability-Review: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Tamir Duberstein <tamird@google.com>