[syslog][rust] Initialize without double-tagging.
fuchsia-syslog always provides a "target" as a tag which
defaults to the module path where the log macro is invoked.
This duplicates the default tag that the syslog library
initializes from the process name. Since we know that we'll
always be providing our own useful tag, pass an empty list of
tags to the underlying library so that it doesn't initialize
with its default process name tag.

Fixed: 55647
Change-Id: I38550de6b6985fae52463c1585e474202cb07bfb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/407999
Reviewed-by: Saman Sami <samans@google.com>
Commit-Queue: Adam Perry <adamperry@google.com>
Testability-Review: Adam Perry <adamperry@google.com>