[bootsvc] Skip unnamed VMOs early
We currently error out late when there's an unnamed VMO passed to us.
Instead, skip the VMO early in processing.

We can receive an unnamed VMO if InstrumentationData::GetVmos() does not
call InstrumentationData::Publish() for a particular VMO. Since we can't
close the handle to the VMO, as there's an expectation of a fixed set of
handles in the userboot message, we must deal with this later in
bootsvc.

Bug: b/159855311
Change-Id: Ibb56d6eb929fed5eb33770187ccab6d4ea565f21
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/408836
Testability-Review: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Tess Eisenberger <teisenbe@google.com>
Commit-Queue: Abdulla Kamar <abdulla@google.com>