[camera] Add linear back-off for sysmem free space probe.
This change adds a linear back-off to the delay with which sysmem is
probed for free space. Previously, sysmem would occasionally not free
deallocated memory within the fixed delay, and so eventually fail to
allocate buffers within the fixed 100ms time. This change increases the
initial delay to 200ms, and increases it each attempt up to a maximum of
1000ms.

Test: Ran camera-gym and observed expected behavior.
Change-Id: I5c98ba1455424ad74f25b23ace4d64529b1dc536
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/409036
Reviewed-by: Ernest Hua <ernesthua@google.com>
Testability-Review: Matt Sandy <msandy@google.com>
Commit-Queue: Matt Sandy <msandy@google.com>